% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clap.R
\name{clap}
\alias{clap}
\title{clap: Detecting Class Overlapping Regions in Multidimensional Dat}
\description{
The issue of overlapping regions in
multidimensional data arises when different classes
or clusters share similar feature representations,
making it challenging to delineate distinct boundaries
between them accurately. This package provides methods
for detecting and visualizing these overlapping
regions using partitional clustering techniques
based on nearest neighbor distances.
}
\author{
Priyanga Dilini Talagala
_PACKAGE
}
